backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pdfminer.six==20221105
spacy==3.7.4
nltk==3.8.1
scikit-learn==1.4.2
numpy==1.26.4
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
anthropic==0.28.0
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.7.1
